Island Survival - 3rd Person Action Game
Island Survival is a 3rd person action game built on Irrlicht 3D Engine
which has 5 waves of enemies which are characterized into 3 types -
Fairie, Monster and Wraith. The player has to defeat all the waves of
enemies in order to finish the game.

P.S. : Sorry for the bad models and textures. I couldn't get better ones
for free...:)